requirements.txt
fastapi
uvicorn[standard]
cryptography
python-dotenv
pymongo[srv]
scikit-learn
